Create seg2io.py
This is an initial move of the `rad2n=py` code from `jeepr`: https://github.com/agilescientific/jeepr/blob/master/jeepr/rad2np.py

It is now two functions, since we want to get an xarray back. This means that we need to have sensible headers. Given that the standard does not specify that we should have particular ones, we can not rely on having them. Settling (for now) on making the user assign names for coords of the `xarray.Dataset`.

This seems to work for the data files that I have access to, but I am looking for more.